// bioskop/src/app/Models/actor.ts
export interface ActorFile {
    id: number;
    name: string;
    dateOfBirth: Date;
    bio: string;
    pictureFile: File;
    pictureString?: string;
    character?: string;
};
export interface ActorStringPROBAJBEZOVOGA {
    id: number;
    name: string;
    dateOfBirth: Date;
    bio: string;
    pictureString: string;
};